% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInformation.R
\docType{data}
\name{CedarRRain}
\alias{CedarRRain}
\title{CedarRRain}
\description{
Instantaneous values of rainfall from Cedar Creek
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{rainmaker}
