% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInformation.R
\docType{data}
\name{cedarSamples}
\alias{cedarSamples}
\title{cedarSamples}
\description{
Beginning and ending dates and times for composite samples collected at Cedar Creek. This information can be used to determine rainfall for specific time periods using RMeventsSamples.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{rainmaker}
