% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_events.R
\name{discharge_events}
\alias{discharge_events}
\title{discharge_events}
\usage{
discharge_events(df, ieHr = 6, qthresh, discharge = "Value",
  time = "pdate")
}
\arguments{
\item{df}{dataframe with discharge}

\item{ieHr}{numeric Interevent period in hours, defaults to 6. 
The amount of time between discharge measurements above threshold required to be considered a new event.}

\item{qthresh}{numeric Discharge threshold value, over which an event is considered to be occuring.}

\item{discharge}{string Column name where discharge values are stored. Defaults to "Value".}

\item{time}{string column name where dates are stored as POSIXct values, defaults to "pdate"}
}
\value{
dataframe of all discharge events based on ieHr and qthresh criteria. Includes start and end times for each event.
}
\description{
Discharge event determination - calculates event start and end times based on discharge record.
}
