% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInformation.R
\docType{data}
\name{precip_05408480}
\alias{precip_05408480}
\title{precip_05408480}
\description{
Precip sample data from site 05408480
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{rainmaker}
