% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInformation.R
\docType{package}
\name{Rainmaker-package}
\alias{Rainmaker-package}
\title{Rainmaker}
\description{
Rainmaker is a collection of functions used to process instantaneous rainfall data to define event rainfall depths, rainfall intensities and antecedent rainfall values. Event definition can be determined solely within rainmaker using the instantaneous rainfall data or can be focused directly on dates and times of interest (for example, using specific water quality sampling periods).
}
\details{
\tabular{ll}{
Package: \tab Rainmaker\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-01-10\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}

Collection of functions..
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
