% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMerosivity.R
\name{RMerosivity}
\alias{RMerosivity}
\title{RMerosivity
This function computes the erosive power of the rainfall}
\usage{
RMerosivity(df, ieHr, timeInterval, method, rain = "rain",
  StormSummary = StormSummary)
}
\arguments{
\item{df}{dataframe with instantaneous rainfall, same df used for RMIntense}

\item{ieHr}{time between events in hours}

\item{timeInterval}{the minimum time between rows of data, 
assumed to be the time it took for precip to accumulate in the collector before the first tip of an event}

\item{method}{choose which energy equation to use (see below)}

\item{rain}{string column name of rainfall unit values, defaults to "rain"}

\item{StormSummary}{dataframe output by RMIntense, defaults to "StormSummary"

method=1: McGregor (1995) Supercedes Brown and Foster equation (1987), which superceded Agriculture Handbook 537 (1979).
method=2: Wischmeier, Agriculture Handbook 537 (1979, 1981), correct computation of formula 2 found in AH537
method=3: Original Rainmaker (1997) USGS Wisconsin Water Science Center, based on equation in Agriculture Handbook 537. Storms with I30>2.5 are incorrectly computed.}
}
\description{
RMerosivity
This function computes the erosive power of the rainfall
}
