% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMevents.plotQ.R
\name{RMevents.plotQ}
\alias{RMevents.plotQ}
\title{RMevents.plotQ}
\usage{
RMevents.plotQ(
  df,
  dfQ,
  date = "pdate",
  Qdate = "pdate",
  rain = "rain",
  Q = "Q",
  df.events,
  sdate = "StartDate",
  edate = "EndDate",
  erain = "depth",
  plot.buffer = 3,
  logy = "",
  site.name = "",
  SampleInfo,
  sampbdate = "",
  sampedate = ""
)
}
\arguments{
\item{df}{dataframe with unit value rainfall data}

\item{dfQ}{dataframe with unit value Q data}

\item{date}{string Date column in df as POSIXct}

\item{Qdate}{string Date column in dfQ as POSIXct}

\item{rain}{string Column in df with instantaneous rain values}

\item{Q}{string Column in dfQ with instantaneous Q values}

\item{df.events}{dataframe with start and end dates/times for events}

\item{sdate}{string Start date column in df.events rain file as POSIXct}

\item{edate}{string End date column in df.events as POSIXct}

\item{erain}{string Event rainfall depth column in df.events}

\item{plot.buffer}{numeric Used to define plotting window in days. Graphs will include}

\item{logy}{string "y" if log y-axis for Q or "" if linear axis. Will default to "".}

\item{site.name}{site name as data type character}

\item{SampleInfo}{if TRUE then sample start and end dates/times are plotted on the hydrograph;
if FALSE then sample start and end dates/times are not plotted on the hydrograph.}

\item{sampbdate}{character column name in df.events for the beginning of the sampling period}

\item{sampedate}{character column name in df.events for the ending of the sampling period}
}
\value{
plots of rainfall events and discharge
}
\description{
Function to graph rainfall and flow for a given x-day window around specified event periods
}
\examples{
#Example 1 - Rainfall/Q plots without sample start/end arrows
RDB <- CedarRRain
dfQ <- cedarq
dfQ <- RMprep(dfQ,prep.type=1,date.type=3,tz="CST6CDT")
RDB2 <- RMprep(RDB,prep.type=1,date.type=1,dates.in="CST.Time",tz="CST6CDT")
RDB3 <- subset(RDB2,upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.>-1)
event.list <- RMevents(df=RDB3,ieHr=6,rainthresh=0.2,rain="upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.")
events.0.2 <- event.list$storms2
site.name <- "Example Site"
SampleInfo <- FALSE
pdf(paste(site.name,"_events.pdf",sep=""))
RMevents.plotQ(RDB3,dfQ,rain="upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.",df.events=events.0.2,erain="rain",
site.name=site.name,SampleInfo=SampleInfo)
shell.exec(paste(site.name,"_events.pdf",sep=""))
dev.off()
#Example 2- Rainfall/Q plots with sample start/end arrows
RDB <- CedarRRain
cedarSamples <- cedarSamples
names(RDB)[2] <- "UVRain"
RDB2 <- RMprep(RDB,prep.type=1,date.type=1,dates.in="CST.Time",tz="CST6CDT")
eventListSamples <- RMeventsSamples(df=RDB2,ieHr=6,rain="UVRain",time="pdate",dfsamples=cedarSamples,bdate="pSstart",edate="pSend")
dfQ <- cedarq
dfQ <- RMprep(dfQ,prep.type=1,date.type=3,tz="CST6CDT")
site.name <- "Example Site"
SampleInfo <- TRUE
sampbdate <- "pSstart"
sampedate <- "pSend"
pdf(paste(site.name,"_events.pdf",sep=""))
RMeventsSamples.plotQ(RDB2,dfQ,rain="UVRain",df.events=eventListSamples,sdate="Braindate",edate="Eraindate",
erain="depth",logy="",site.name=site.name,sampbdate="pSstart",sampedate="pSend")
shell.exec(paste(site.name,"_events.pdf",sep=""))
dev.off()
}
