% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMevents_sample.R
\name{RMevents_sample}
\alias{RMevents_sample}
\title{RMevents_sample}
\usage{
RMevents_sample(df, ieHr = 6, rain = "rain", time = "pdate",
  dfsamples, bdate = "bpdate", edate = "epdate")
}
\arguments{
\item{df}{dataframe with rainfall}

\item{ieHr}{numeric Interevent period in hours, defaults to 6,}

\item{rain}{string Column name of rainfall unit values, defaults to "rain"}

\item{time}{string column with as.POSIXctdate, defaults to "pdate"}

\item{dfsamples}{dataframe with the beginning and ending dates and times
of sampling periods in POSIXct format}

\item{bdate}{character column name in dfsamples for the beginning of the sampling period}

\item{edate}{character column name in dfsamples for the ending of the sampling period}
}
\value{
list of storms and storms2
}
\description{
Compute rainfall event variables based on time series of rain data with only 
one rain gage or one mean radar rain column. The function does not calculate events based on
the rain data itself (such as in RMevents), but rather calculates event variables based on an input
of sample/event start and end times.
}
\examples{
RDB <- CedarRRain
cedarSamples <- cedarSamples
names(RDB)[2] <- "UVRain"
RDB2 <- RMprep(RDB,prep.type=1,date.type=1,
               dates.in="CST.Time",tz="CST6CDT")
eventListSamples <- RMeventsSamples(df=RDB2,ieHr=6,
                                    rain="UVRain",
                                    time="pdate",
                                    dfsamples=cedarSamples,
                                    bdate="pSstart",edate="pSend")
}
