% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMintensity.R
\name{RMintensity}
\alias{RMintensity}
\title{RMintensity}
\usage{
RMintensity(df, date = "r.date", rain = "rain", df.events,
  sdate = "StartDate", edate = "EndDate", depth = "depth", xmin = c(60,
  180, 360))
}
\arguments{
\item{df}{dataframe}

\item{date}{string Date column name in df as POSIX}

\item{rain}{string Column name in df with instantaneous rain values}

\item{df.events}{dateframe with start and end dates/times for events}

\item{sdate}{string Start date column in df.events rain file as POSIX}

\item{edate}{string End date column in df.events rain file as POSIX}

\item{depth}{string rain depth in event file, defaults to "depth",}

\item{xmin}{vector vector of values representing X-minute max rainfall requested}
}
\value{
df.events dataframe, X-hour maximum rainfall intensities
}
\description{
Function to compute maximum x-minute rainfall intensities in units of depth/hr
}
\examples{
RDB <- CedarRRain
RDB2 <- RMprep(RDB,prep.type=1,date.type=1,dates.in="CST.Time",tz="CST6CDT")
RDB3 <- subset(RDB2,upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.>-1)
event.list <- RMevents(df=RDB3,ieHr=6,rainthresh=0.2,rain="upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.")
events.0.2 <- event.list$storms2
intensities <- RMIntense(RDB3,date="pdate",rain="upload.ph3_site_basin_cedar_creek.Id.0....Geographical.Mean.kg.m.2.",events.0.2,sdate="StartDate",edate="EndDate",depth="rain",xmin=c(5,15,30))
}
